+++ gulpfile.js
@@ -48,5 +48,4 @@ gulp.task('watch', function () {
   gulp.watch(['generators/**/*.js', 'test/**'], ['test']);
 });
 
-gulp.task('prepublish', ['nsp']);
 gulp.task('default', ['static', 'test']);
