+++ package.json
@@ -56,7 +56,6 @@
   },
   "repository": "elasticio/generator-elasticio",
   "scripts": {
-    "prepublish": "gulp prepublish",
     "test": "gulp"
   },
   "license": "Apache-2.0"
